+++ app/Http/Controllers/DomicilioController.php
@@ -21,6 +21,7 @@ class DomicilioController extends Controller
      */
     public function index()
     {
+        $this->authorize('ver-domicilios');
         $domicilios = Domicilio::orderBy('departamento')->orderBy('calle')->paginate(15);
         return view('domicilios.index', compact('domicilios'));
     }
@@ -30,6 +31,7 @@ class DomicilioController extends Controller
      */
     public function create()
     {
+        $this->authorize('gestionar-domicilios');
         return view('domicilios.create');
     }
 
@@ -38,6 +40,7 @@ class DomicilioController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('gestionar-domicilios');
         $validated = $request->validate([
             'departamento' => 'required|string|max:100',
             'provincia' => 'nullable|string|max:100',
@@ -66,6 +69,7 @@ class DomicilioController extends Controller
      */
     public function show(Domicilio $domicilio)
     {
+        $this->authorize('ver-domicilios');
         $domicilio->load('procedimientos');
         return view('domicilios.show', compact('domicilio'));
     }
@@ -75,6 +79,7 @@ class DomicilioController extends Controller
      */
     public function edit(Domicilio $domicilio)
     {
+        $this->authorize('gestionar-domicilios');
         return view('domicilios.edit', compact('domicilio'));
     }
 
@@ -83,6 +88,7 @@ class DomicilioController extends Controller
      */
     public function update(Request $request, Domicilio $domicilio)
     {
+        $this->authorize('gestionar-domicilios');
         $validated = $request->validate([
             'departamento' => 'required|string|max:100',
             'provincia' => 'nullable|string|max:100',
@@ -111,6 +117,7 @@ class DomicilioController extends Controller
      */
     public function destroy(Domicilio $domicilio)
     {
+        $this->authorize('gestionar-domicilios');
         $domicilio->delete();
 
         return redirect()->route('domicilios.index')
